refactor: move page files
