+++ hk11/lib/pages/login_page.dart
@@ -1,8 +1,8 @@
-import 'package:hk11/pages/login_page/profile_page.dart';
+import 'package:hk11/pages/profile_page.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:hk11/pages/login_page/onboarding.dart';
+import 'package:hk11/pages/onboarding.dart';
 import 'view.dart';
 
 class LoginPage extends StatefulWidget {
